StringUtilities: Added lastIndexOf function for type String (native one is for char only). Added isLetterOrDigit function. Utilities: Made getNonZeroBytes functions first letter lower case. Moved ToArray to the new CollectionItilites class and made it's first letter lower case.
Additions:
CollectionUtilities:
Vector and Hashtable "equals" functions that do shallow comparisions of the objects.
toArray, from Utilities.
putAll to copy a Hashtable or Vector to another Hashtable or Vector.
Synchronized and readonly Vector and Hashtable functions.
NIO:
The 5.0 and higher NIO classes, a subset of them, avaliable for pre-5.0 devices.